% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCtransform.R
\name{ABCtransform}
\alias{ABCtransform}
\title{Adaptive Box-Cox transformation}
\usage{
ABCtransform(
  FeatureTable,
  Impt = "default",
  LambdaRange = c(-3, 3),
  GapIdentifier = 0,
  SampleInCol = TRUE,
  Output = FALSE
)
}
\arguments{
\item{FeatureTable}{Data frame with features in row and samples in column (default).}

\item{Impt}{A single string specifying the imputation method to be used.}

\item{LambdaRange}{A numeric vector indicating the optimization range of lambda value.}

\item{GapIdentifier}{A single value to identify gaps in \code{FeatureTable}.}

\item{SampleInCol}{\code{TRUE} if samples are in column. \code{FALSE} if samples are in row.}

\item{Output}{\code{TRUE} will output the result table in the current working directory.}
}
\value{
This function returns the transformed data frame.
}
\description{
Feature-specific data transformation to improve data normality in untargeted metabolomics.
}
\details{
Adaptive Box-Cox (ABC) transformation is designed to improve the data normality in untargeted metabolomics.
ABC transformation contains three modules: \cr
1. Data pre-treatment by gap-filling and data scaling \cr
2. Optimize lambda value for ABC transformation \cr
3. Perform the ABC data transformation using the optimized lambda value \cr

\code{FeatureTable} contains measured signal intensities of metabolic features,
with features in row and samples in column (default). The column names should
be sample names, and the first row should be sample group names (e.g. control, case).\cr

Four imputation methods are provided here: \cr
1. \code{default}, default imputation method by ABC transformation \cr
2. \code{knn}, KNN method supported by \pkg{VIM} package. See \code{\link[VIM]{kNN}} for details. \cr
3. \code{addition}, only replace gaps with identical small values (min/5). \cr
4. \code{rf}, random forest method supported by \pkg{mice} package. See \code{\link[VIM]{mice}} for details.
This method is not recommended due to long calculation time when feature number > 200. \cr
}
\examples{
TransformedTable = ABCtransform(DemoData)
}
\references{
To be updated.
}
