% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnn_delta_tanh}
\alias{dnn_delta_tanh}
\title{Backprop subroutine for delta}
\usage{
dnn_delta_tanh(W, H, delta)
}
\arguments{
\item{W}{weight matrix}

\item{H}{H matrix}

\item{delta}{Input matrix}
}
\description{
Backprop subroutine for delta
}

