% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnn_grad}
\alias{dnn_grad}
\title{Comput the gradient given a DNN and input matrix x, target y}
\usage{
dnn_grad(Wbs, x, y)
}
\arguments{
\item{Wbs}{Weight matrics of DNN, in a list}

\item{x}{Input, n x p matrix}
}
\description{
Comput the gradient given a DNN and input matrix x, target y
}

