% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slow.R
\name{dnn_grad_slow}
\alias{dnn_grad_slow}
\title{Compute gradient for DNN}
\usage{
dnn_grad_slow(Ws, bs, x, y)
}
\arguments{
\item{Ws}{Weights of DNN}

\item{bs}{Bias terms of DNN}

\item{x}{Training points, dimension n x p_x}

\item{y}{Training targets, dimension n x p_y}
}
\value{
A list containing dWs and dbs
}
\description{
Compute gradient for DNN
}

