% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slow.R
\name{dnn_init_slow}
\alias{dnn_init_slow}
\title{Initialize DNN}
\usage{
dnn_init_slow(dims, sigma0 = 0.1, rseed = 315, merge = FALSE)
}
\arguments{
\item{dims}{Size of each layer}

\item{sigma0}{Standard deviation of initial weights}

\item{rseed}{Random seed}
}
\value{
A list containing weight matrices Ws abd  biases bs
}
\description{
Create weight matrices for deep neural network
}

