% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnn_lin}
\alias{dnn_lin}
\title{Apply linear transform}
\usage{
dnn_lin(W, b, x)
}
\arguments{
\item{W}{weight matrix}

\item{b}{bias matrix}

\item{x}{Input, n x p matrix}
}
\description{
Apply linear transform
}

