% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slow.R
\name{dnn_loss_slow}
\alias{dnn_loss_slow}
\title{Computes the loss}
\usage{
dnn_loss_slow(Ws, bs, x, y)
}
\arguments{
\item{Ws}{Weights of DNN}

\item{bs}{Bias terms of DNN}

\item{x}{Training points, dimension n x p_x}

\item{y}{Training targets, dimension n x p_y}
}
\value{
Objective function, sqrt(sum(y - hat(y))^2/n)
}
\description{
Computes the loss
}

