% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slow.R
\name{dnn_make_hs}
\alias{dnn_make_hs}
\title{Used for interfacing with Rcpp}
\usage{
dnn_make_hs(Ws, bs, x)
}
\arguments{
\item{Ws}{Weight matrices of DNN}

\item{bs}{Bias terms of DNN}

\item{x}{New points, dimension n x p}
}
\value{
An list of Hs
}
\description{
Creates list needed to store intermediate results
}

