% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnn_predict}
\alias{dnn_predict}
\title{Predict the output given a DNN and input matrix x}
\usage{
dnn_predict(Ws, x)
}
\arguments{
\item{Ws}{Weight matrics of DNN, in a list}
}
\description{
Predict the output given a DNN and input matrix x
}

