% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/par.R
\name{dnn_sgd_par}
\alias{dnn_sgd_par}
\title{Optimizes the network using sgd}
\usage{
dnn_sgd_par(Wbs, x_tr, y_tr, x_te, y_te, epoch = 100, reg = 0,
  minibatch = 25, alpha0 = 0.1, mc.cores = 4)
}
\arguments{
\item{Wbs}{Weights and biases of DNN}

\item{x_tr}{Training points, dimension n x p_x}

\item{y_tr}{Training targets, dimension n x p_y}

\item{x_te}{Test points, dimension n x p_x}

\item{y_te}{Test targets, dimension n x p_y}

\item{epoch}{Number of epochs}

\item{reg}{Weight regularization}

\item{minibatch}{Minibatch size}

\item{alpha0}{Initial step size}

\item{mc.cores}{MC cores for parallel}
}
\value{
Results
}
\description{
Optimizes the network using sgd
}

