% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slow.R
\name{dnn_sgd_slow}
\alias{dnn_sgd_slow}
\title{Optimizes the network using sgd}
\usage{
dnn_sgd_slow(Ws, bs, x_tr, y_tr, x_te, y_te, epoch = 100, reg = 0,
  minibatch = 100, alpha0 = 0.1)
}
\arguments{
\item{Ws}{Weights of DNN}

\item{bs}{Bias terms of DNN}

\item{x_tr}{Training points, dimension n x p_x}

\item{y_tr}{Training targets, dimension n x p_y}

\item{x_te}{Test points, dimension n x p_x}

\item{y_te}{Test targets, dimension n x p_y}

\item{epoch}{Number of epochs}

\item{reg}{Weight regularization}

\item{minibatch}{Minibatch size}

\item{alpha0}{Initial step size}
}
\value{
Objective function, sqrt(sum(y - hat(y))^2/n)
}
\description{
Optimizes the network using sgd
}

