% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnn_tanh}
\alias{dnn_tanh}
\title{Apply tanh function}
\usage{
dnn_tanh(W, b, x)
}
\arguments{
\item{W}{weight matrix}

\item{b}{bias matrix}

\item{x}{Input, n x p matrix}
}
\description{
Apply tanh function
}

