% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_mat}
\alias{extract_mat}
\title{Extract arma mat from list}
\usage{
extract_mat(ll, i)
}
\arguments{
\item{ll}{List of NumericMatrix}

\item{i}{Index to extract}
}
\description{
Extract arma mat from list
}

