% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{Imputation}
\alias{Imputation}
\title{Data imputation}
\usage{
Imputation(
  FeatureTable,
  Impt = "default",
  GapIdentifier = 0,
  SampleInCol = TRUE,
  Output = FALSE
)
}
\arguments{
\item{FeatureTable}{Data frame with features in row and samples in column (default).}

\item{Impt}{A single string specifying the imputation method to be used.}

\item{GapIdentifier}{A numeric vector indicating the optimization range of lambda value.}

\item{SampleInCol}{\code{TRUE} if samples are in column. \code{FALSE} if samples are in row.}

\item{Output}{\code{TRUE} will output the result table in the current working directory.}
}
\value{
This function returns the imputed data frame.
}
\description{
Missing value imputation by different methods.
}
\details{
Three imputation methods are provided here: \cr
1. \code{default}, default imputation method by ABC transformation \cr
2. \code{knn}, KNN method supported by \pkg{VIM} package. See \code{\link[VIM]{kNN}} for details. \cr
3. \code{addition}, only replace gaps with identical small values (max/1000). \cr
4. \code{rf}, KNN method supported by \pkg{mice} package. See \code{\link[VIM]{mice}} for details. \cr

\code{FeatureTable} contains measured signal intensities of metabolic features,
with features in row and samples in column (default). The column names should
be sample names, and the first row should be sample group names (e.g. control, case).\cr
}
\examples{
ImputedTable = Imputation(DemoData)

}
\references{
To be updated.
}
