% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ising_chain}
\alias{ising_chain}
\title{Runs MCMC on an Ising matrix}
\usage{
ising_chain(x, theta, nits)
}
\arguments{
\item{x}{Ising matrix}

\item{theta}{Ising parameter}

\item{nits}{Number of iterations}
}
\description{
Runs MCMC on an Ising matrix
}

