% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_sign_mat}
\alias{random_sign_mat}
\title{Generates binomial random matrix of signs}
\usage{
random_sign_mat(n)
}
\arguments{
\item{n}{dim}
}
\description{
Generates binomial random matrix of signs
}

