% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{zattach}
\alias{zattach}
\title{Loads all elements of a list into the global environment}
\usage{
zattach(ll)
}
\arguments{
\item{ll}{List with named elements}
}
\value{
No return: side effect is to load elements into environment,
similar to attach() but into environment rather than search path
}
\description{
Loads all elements of a list into the global environment
}

