% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slow.R
\name{dnn_grad_slow}
\alias{dnn_grad_slow}
\title{Compute gradient for DNN}
\usage{
dnn_grad_slow(Ws, x, y)
}
\arguments{
\item{Ws}{Weights of DNN}

\item{x}{Training points, dimension n x p_x}

\item{y}{Training targets, dimension n x p_y}
}
\value{
A list of gradients
}
\description{
Compute gradient for DNN
}

