% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/slow.R
\name{dnn_predict_slow}
\alias{dnn_predict_slow}
\title{Predict given DNN and tanh transfer function}
\usage{
dnn_predict_slow(Ws, bs, x)
}
\arguments{
\item{Ws}{Weight matrices of DNN}

\item{bs}{Bias terms of DNN}

\item{x}{New points, dimension n x p}
}
\value{
An n x ? matrix of predictions
}
\description{
Predict given DNN and tanh transfer function
}

